Refactoring SketchPreview into SketchForm
Removing SketchPreview from SketchCreate/Update, moving to SketchForm.
Minor CSS styling in index.scss.

Note: there is a known bug when accepting user input in the 'Meter'
category of SketchForm. Currently, there is no option to select 'none'
and a user must input a time signature. This will be addressed in
future commits.